refactor(linter): split markdown rules domain (Task 2.5)
- split 204 lines of markdown.rs into logical submodules
- extract discovery, structure modules
- establish markdown orchestration module
